reorganized components and pages
