+++ package.json
@@ -16,7 +16,8 @@
     "react-icons": "^3.11.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.4",
-    "styled-components": "^5.2.0"
+    "styled-components": "^5.2.0",
+    "webfontloader": "^1.6.28"
   },
   "scripts": {
     "start": "react-scripts start",
