+++ src/Components/ProductDetails/Productdetails.css
@@ -30,9 +30,9 @@
 }
 
 /* this CSS is just to identify the items - it will be changed */
-div{
+/* div{
   border-bottom: solid 0.5px lightgray;
-}
+} */
 
 .secondary-info{
   padding: 10px;
