+++ src/Components/ProductInformation/Productinfo.css
@@ -35,6 +35,6 @@
 }
 
 .product-card p{
-  color: var(--);
+  color: black;
   font-size: 1.8em;
 }
