+++ src/pages/ExplorePage.jsx
@@ -1,12 +1,15 @@
 import React, { Component } from "react";
-import Categories from './Components/CategoriesButtons/ExploreCategories'
-import Footer from './Components/Footer/Footer'
+import { FaIcons } from "react-icons/fa";
+import Categories from '../components/CategoriesButtons/CategoriesButtons'
+import Footer from '../components/Footer/Footer'
+import '../index.css'
 
 function ExplorePage() {
   return (
     <div>
       <div>
         <h1>Explore Page</h1>
+        <input className="search-bar" type="text" placeholder="Search"></input>
         <Categories/>
       </div>
       <Footer/>
